Refatora o botão de baixar o curriculo e o separa por ser um componente reutilizavel; Separa tarefas da lista de projetos
