+++ frontend/src/components/common/Projects.jsx
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types';
 import { useContext } from 'react';
 import AppContext from '../../contexts/AppContext';
+import ListProjects from './ListProjects';
 
 function Projects(props) {
   const {
@@ -20,38 +21,11 @@ function Projects(props) {
         <hr className={ `Emphasis-title-${platform}` } />
       </div>
       <ul className={ `Projects-${platform}` }>
-        {
-          dataList.map((item, i) => (
-            <li key={ i } className={ `Project-${modeThame}-${platform}` }>
-              <div className={ `Project-description-${platform}` }>
-                <div
-                  className={ `Project-Bunner-${platform}` }
-                  style={ { backgroundImage: `url(${item.img})` } }
-                >
-                  <a
-                    href={ item.link }
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    Saiba mais
-                    <p>Github</p>
-                  </a>
-                </div>
-                <h3>{ item.description }</h3>
-                <ul className={ `Project-technologies-${platform}` }>
-                  {
-                    item.technologies.map((tec, index) => (
-                      <li key={ index }>
-                        { tec }
-                      </li>
-                    ))
-                  }
-                </ul>
-              </div>
-              <h3 className={ `Project-Tags-${platform}` }>{ item.tags }</h3>
-            </li>
-          ))
-        }
+        <ListProjects
+          platform={ platform }
+          modeThame={ modeThame }
+          dataList={ dataList }
+        />
       </ul>
       <p className={ `Content-description-${platform}` }>
         { description.projectThow }
