+++ frontend/src/layouts/desktop/Projects.css
@@ -118,26 +118,3 @@
     padding-right: 50px;
   }
 }
-
-/* .Project-1 {
-  box-shadow: 0px 0px 10px 1px rgba(0, 0, 0, 0.555);
-}
-
-.Project-1:hover {
-  box-shadow: 0px 0px 20px 2px rgba(0, 0, 0, 0.555);
-  opacity: 1;
-  transform: scale(1.02);
-  font-weight: 600;
-}
-
-.Project-0 {
-  box-shadow: 0px 0px 10px 1px rgba(255, 255, 255, 0.541);
-}
-
-.Project-0:hover {
-  box-shadow:
-    0px 0px 20px 2px rgba(255, 255, 255, 0.836);
-  opacity: 1;
-  transform: scale(1.02);
-  font-weight: 600;
-} */
