+++ frontend/src/layouts/mobile/Footer.css
